+++ src/kv.c
@@ -2,7 +2,7 @@
 #include "kv.h"
 
 
-kv_t* kv_init(int capacity){
+kv_t* kv_init(size_t capacity){
 
     if (capacity <= 0) return NULL;
